refractor: Move everything to the darkcitizens package
