+++ src/main/kotlin/com/github/recraftedcivilizations/darkcitizens/dPlayer/DPlayerFactory.kt
@@ -1,6 +1,5 @@
-package com.github.recraftedcivilizations.dPlayer
+package com.github.recraftedcivilizations.darkcitizens.dPlayer
 
-import com.github.recraftedcivilizations.jobs.IJob
 import org.bukkit.entity.Player
 import java.util.*
 
@@ -18,7 +17,7 @@ object DPlayerFactory {
      * For the arguments take a look at DPlayerData1
      * @see [DPlayerData1]
      */
-    fun createDPlayer(uuid: UUID, job: String?, wanted: Boolean,isCriminal: Boolean, groupLvls: Map<String, Int>, groupXps: Map<String, Int>): DPlayer{
+    fun createDPlayer(uuid: UUID, job: String?, wanted: Boolean,isCriminal: Boolean, groupLvls: Map<String, Int>, groupXps: Map<String, Int>): DPlayer {
         val dData = DPlayerData1(uuid, job, wanted ,isCriminal , groupLvls , groupXps)
         return DPlayer(dData)
     }
@@ -28,7 +27,7 @@ object DPlayerFactory {
      * For the arguments take a look at DPlayerData2
      * @see [DPlayerData2]
      */
-    fun createDPlayer(job: String?, wanted: Boolean,isCriminal: Boolean, groupLvls: Map<String, Int>, groupXps: Map<String, Int>, player: Player): DPlayer{
+    fun createDPlayer(job: String?, wanted: Boolean,isCriminal: Boolean, groupLvls: Map<String, Int>, groupXps: Map<String, Int>, player: Player): DPlayer {
         val dData = DPlayerData2(job, wanted ,isCriminal , groupLvls , groupXps)
         return DPlayer(dData, player)
     }
@@ -37,7 +36,7 @@ object DPlayerFactory {
      * Create a new empty DPlayer based on a player
      * @param player The player to create the DPlayer for
      */
-    fun createDPlayer(player: Player): DPlayer{
+    fun createDPlayer(player: Player): DPlayer {
         return DPlayer(player)
     }
 }
