+++ src/main/kotlin/com/github/recraftedcivilizations/darkcitizens/jobs/IJob.kt
@@ -1,7 +1,7 @@
-package com.github.recraftedcivilizations.jobs
+package com.github.recraftedcivilizations.darkcitizens.jobs
 
-import com.github.recraftedcivilizations.dPlayer.DPlayer
-import com.github.recraftedcivilizations.tasks.ITask
+import com.github.recraftedcivilizations.darkcitizens.recraftedcivilizations.dPlayer.DPlayer
+import com.github.recraftedcivilizations.darkcitizens.recraftedcivilizations.tasks.ITask
 import org.bukkit.entity.Player
 import java.util.*
