+++ src/main/kotlin/com/github/recraftedcivilizations/darkcitizens/tasks/ITask.kt
@@ -1,7 +1,7 @@
-package com.github.recraftedcivilizations.tasks
+package com.github.recraftedcivilizations.darkcitizens.tasks
 
-import com.github.recraftedcivilizations.dPlayer.DPlayer
-import com.github.recraftedcivilizations.tasks.actions.IAction
+import com.github.recraftedcivilizations.darkcitizens.recraftedcivilizations.dPlayer.DPlayer
+import com.github.recraftedcivilizations.darkcitizens.recraftedcivilizations.tasks.actions.IAction
 import org.bukkit.entity.Player
 
 /**
