+++ src/main/kotlin/com/github/recraftedcivilizations/darkcitizens/tasks/Task.kt
@@ -1,10 +1,10 @@
-package com.github.recraftedcivilizations.tasks
+package com.github.recraftedcivilizations.darkcitizens.tasks
 
-import com.github.recraftedcivilizations.BukkitWrapper
-import com.github.recraftedcivilizations.dPlayer.DPlayer
-import com.github.recraftedcivilizations.dPlayer.DPlayerManager
-import com.github.recraftedcivilizations.jobs.JobManager
-import com.github.recraftedcivilizations.tasks.actions.IAction
+import com.github.recraftedcivilizations.darkcitizens.BukkitWrapper
+import com.github.recraftedcivilizations.darkcitizens.recraftedcivilizations.dPlayer.DPlayer
+import com.github.recraftedcivilizations.darkcitizens.recraftedcivilizations.dPlayer.DPlayerManager
+import com.github.recraftedcivilizations.darkcitizens.recraftedcivilizations.jobs.JobManager
+import com.github.recraftedcivilizations.darkcitizens.recraftedcivilizations.tasks.actions.IAction
 import net.milkbowl.vault.economy.Economy
 import org.bukkit.boss.BarColor
 import org.bukkit.boss.BarStyle
@@ -30,7 +30,7 @@ class Task(
     private val dPlayerManager: DPlayerManager,
     private val econ: Economy,
     private val jobManager: JobManager,
-    private val bukkitWrapper: BukkitWrapper = BukkitWrapper()
+    private val bukkitWrapper: com.github.recraftedcivilizations.darkcitizens.BukkitWrapper = com.github.recraftedcivilizations.darkcitizens.BukkitWrapper()
 ) : ITask {
 
     /**
